[BTLS]: Initial import of the BTLS sources.
This is copied from the 'martin-btls-stable' branch, commit
cda0c2528dcca5dcbf8019d114eff804e953e619.

There is a bunch of testing and debugging code in here which still
needs to be removed and cleaned up, this is just the initial import
to get things to build on the bots.