[runtime] Move MonoClass::generic_class to MonoClassGenericInst.
To make this possible, two accessor functions were added/augmented:

mono_class_get_generic_class: Gets the generic_class and aborts if the MonoClass is of the wrong kind.
mono_class_try_get_generic_class: Gets the generic_class or return NULL if the MonoClass is of the wrong kind.

The bulk of this change is converting all checks for generic_class into calls to mono_class_is_ginst.