[marshal] Add MarshalAs(UTF8) support and fix UTF8 marshalling
This adds support for the new UnmanagedType.LPUTF8Str marshaling encoding,
which allows developers to flag strings and stringbuilders to be encoded
to and from using UTF8.

In Mono, this is mostly a pass-through as we have historically only done
UTF8, so this merely adds the constant processing to the runtime.

FIX: while bringing the .NET Core tests over to Mono, I found that our
utf8 encoding made an incorrect assumption when copying data from managed
code to unmanaged code: we only copied the number of bytes equivalent to
the number of UTF-16 characters, which means that we marshal fewer bytes
than we need when passing strings that contain characters that use more
than one byte.   The same problem happens on the return value, we were copying
fewer bytes than needed.