Handle GetHostByName without a network connection
GetHostByName should still work for localhost even when there is no
network (e.g. the network cable is unplugged). It should return the
proper host name. Since we started supporting IPv6, this was broken.

The getaddrinfo function returns a NULL addrinfo value when there is no
network. If that happens, fall back to the IPv4-only implementation to
try to resolve the host name.

This corrects Unity case 840534.