[System] Remove any CFNetwork usage from the watchOS profile. Fixes #45847.
The MacProxy class uses CFNetwork, but since CFNetwork is not a public
framework on watchOS, we can't use it.

So remove MacProxy completely (it only contains internal classes), and throw
PlatformNotSupportedException in any API that used it (the managed networking
stack is not supported on watchOS anyway, so this should be safe).

This is a backport of the following commits:
c023b2b30aa3bd0d2f0eafb5e983861df7f67de4 and
16b8979fd1698669499defc95cc069c0adf1e291.

https://bugzilla.xamarin.com/show_bug.cgi?id=45847