[mono-threads] Fix unified thread suspend
`mono_threads_suspend_begin_async_suspend` return FALSE if it coudn't successfully suspend the thread, not if the thread is not in the correct state. The latter is checked by calling `mono_threads_suspend_check_suspend_result` which will return FALSE if we couldn't initialize the state of the suspended thread.