[gc] Remove mono_gc_set_stack_end (#3845)
* [sgen] Add stress test

This hope to suspend a thread while it is in the process of attaching or detaching.

* [sgen] Fix unified suspend

This would cause a crash with unified suspend: if a thread would be suspended at mono-threads.c:1168, the stack_start would be greater than the stack_end, triggering the g_error at sgen-stw.c:351. The stack_end would initially be initialized in `sgen_client_thread_register`, but it would be overriden in start_wrapper at threads.c:893.