[threads] Use MonoOSEvent for Thread.Suspend/Resume
By using a MonoOSEvent we avoid relying on the utils/threads suspend mechanism to suspend a metadata/threads. This reduce coupling between both mechanism, and that would allow us to remove a bunch of code from utils/threads when switching to coop.

This only works with coop suspend, as we need the ability to return from an async call, and that's not possible with preemptive suspend.