[corlib] Expose Marshal.PtrToStringBSTR on mobile
The SecureStringHelper from referencesource I added in #3887
relies on this to convert from SecureString to plaintext.

It failed on mobile because we always throw a NotImplementedException there,
but we actually have an implementation on the runtime side (even in the DISABLE_COM case).

(cherry-picked from df1475d95e1dd4327b44df459d1aabbecb8ce92f)