Fixes bugs in System.ServiceProcess on Windows
There was a crash on Windows x64 in Win32ServiceController when enumerating
devices and services. Also fixed the ServiceControllerTest to succeed on .NET
as well Mono on Windows. Had to replace the services used by the test since
most of them have been removed from recent Windows versions.

This PR also enables the System.ServiceProcess tests in Windows PR builds.