[ci] Allow skipping runtime/classlib test suites via GitHub label
Inspired by the xamarin-macios team, they use GitHub labels for
telling Jenkins to run specific tests on PRs.

For us it probably makes more sense to do it the other way round
and exclude either the runtime tests or the class library
tests from running.

When I'm making a bugfix in the class libs the chances of it
breaking the runtime tests are quite low, yet I still have to
wait ~1hr before Jenkins even starts running the class lib tests.

The label check happens before the test suites, so with this I'd
have about 30mins to decide whether I want to skip and add the label
on GitHub.