Don't allow Join to return due to an APC (#3970)
* Don't allow Join to return due to an APC

On Windows, Thread.Join does an alertable wait. If the thread calling
Join gets an APC during the call, Join will incorrectly return. Instead,
the code should handle the APC and continue waiting.

* Do an uninterrupted wait during thread join.

Mimic the behavior of mono_wait_uninterrupted, but for a single thread
that we want to join.

* Make some style changes suggested in the PR.