[w32handle] Fix race in foreach and unref
It could happen that we unref a handle that we are currently looping over. This could be the case if we are calling Process.GetProcessById (which loops over all handles) and mono_w32handle_unref on a MonoW32HandleProcess handle.

We need to make the `scan_mutex` and recursive mutex, as we could call `mono_w32handle_unref` which could destroy the handle, while in `mono_w32handle_foreach`.