[mkbundle] Don't use Path.GetDirectoryName on `directory` (#4130)
`directory` is already a directory. It's in the name.

`Path.GetDirectoryName ("/path/to/a/thing")` returns `"/path/to/a"`, whether
`thing` is a directory or not. If there's a trailing `/` then it returns
the directory, not the parent - but we don't have a trailing `/`.

This fixes an issue where the path needs to be created by hand before an
SDK can be downloaded.
(cherry picked from commit 031d79c2c4ed70ee9675c53fab676a2ddaca0922)